refactor(#113926): Extracting "JsonValueOfJsonBool" to separate file.
